Configures the Blazor app for Docker deployment
Configures the Blazor application to run within a Docker container.
Adds a Dockerfile for building the Blazor application image.
Updates the docker-compose files to orchestrate the Blazor container.
Configures Kestrel to use HTTPS with a default certificate in Docker.
Updates client projects to point to the correct localhost ports in Docker.